Pass forAccounts info through correctly
forAccounts determines, at the origin point of a `transaction` event,
which accounts that are being monitored are involved and therefore
should reflect the transaction. This is also used downstream to do
things like mark those accounts for balance checks.

This info was being reconstructed in `Main` irrespective of monitoring
status for unclear reasons, and lost entirely when a transaction was
enriched. Now, the original chain-service `forAccounts` is preserved
when passing along to Redux, and enriched transactions carry the
original information as well.

Future work could use transaction enrichment to further enrich
`forAccounts` with additional accounts affected by the transaction.